requirements.txt
numpy>=1.21.0
pandas==1.4.3
vcrpy>=4.1.1
requests>=2.27.1
ipykernel
jupyterlab-myst
matplotlib
itables
nbconvert
python-dotenv